factory/src/main/java/App.java

public class App {

    private King king;
    private Castle castle;
    private Army army;

    public App(King king) {
    }


    public void createKingdom(final KingdomFactory factory) {
        this.king = factory.createKing();

    }

    public String getDescription() {
        return this.king.getDescription();
    }

    King getKing(final KingdomFactory factory) {
        return factory.createKing();
    }

    public King getKing() {
        return king;
    }

    private void setKing(final King king) {
        this.king = king;
    }

    Castle getCastle(final KingdomFactory factory) {
        return factory.createCastle();
    }

    public Castle getCastle() {
        return castle;
    }

    private void setCastle(final Castle castle) {
        this.castle = castle;
    }

    Army getArmy(final KingdomFactory factory) {
        return factory.createArmy();
    }

    public Army getArmy() {
        return army;
    }

    private void setArmy(final Army army) {
        this.army = army;
    }


}
